src/main/java/com/example/web/BlogProperties.java
package com.example.web;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;


@Component
public class BlogProperties {

    @Value("${com.colin.name}")
    public String name;

    @Value("${com.colin.age}")
    public int age;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public int getAge() {
        return age;
    }

    public void setAge(int age) {
        this.age = age;
    }
}

